PORTAL-261 Implement drag'n'drop job submission
Using HTML5 File API and DND API, we can now submit jobs using drag and drop.
The drag and drop zone is the top menu bar. The submit button is activated when
dragging a file over it.
It should work with Chrome, Firefox (expect Linux) and IE10.

The lib-gwt-file library was used. It required an upgrad to GWT 2.5.

When dropping a file, the file is sent to the upload form via a hidden field.
It is base64 encoded to be able to used as multipart data.

The root layout has to be changed to a pure GWT layout as I managed to get drag
and drop events to work only with a pure GWT panel. The layouts were not rendering
correctly with a mix of GWT and SmartGWT widgets.

Change-Id: Ibb03565572e86c593415ee929f8a71030b4fdaad