PORTAL-265 Progress bar in job list is not displayed
The listgridrecord seems to recreated on the JS side when adding data to the
datastore (addData call), therefore only attributes set via the setAttribute()
method were retrieved. The job field was not present. So I set the job using
the setAttribute method.
It could be a change in SmartGWT as we upgraded recently.
See http://forums.smartclient.com/showthread.php?t=18457

As the record will then always be a listgridrecord with an job attribute
we can query this attribute directly and do not use the JobRecord instance or
the submittingJobs map.

Change-Id: I5008ed9267e4db83f0a4063135d09ad1698e2db0