PORTAL-280 - Monitoring charts are not displayed immediately
Force two retrievals of data before the timer is activated so the charts
displays something once a node or a host is selected.
The values will probably be very close but the user experience is better as it
does not look like it is slow or stuck.

For the CPU chart, the first value was only saved for latter relative computations,
it is now used and a "before-first" value is extrapolated for the first value to
allow relative computations. This way, the CPU chart has the same number of points
as the others.

Change-Id: Ibf7460df3f5e14eba034316e4d853a15eec18181