PORTAL-284 The tab of monitoring for hosts' processes shows some blank fields.
ProcessesView queries through JMX the MBean sigar:Type=Processes from
one RMNode in the monitored host. The result of the query is a json
array whose elements are maps (one map per process) with property:value
(for instance "pid":1022). Until now, for every property foundi for a
process it was mandatory to have an existing column in the ListGrid
where to display the value of the process' property. Now if this
row does not exist the property's value will be just not shown.

Change-Id: I5ea1bf68c23bca08dfe29070c9b95ebae7509e19
Signed-off-by: Mauricio Jost <mauricio.jost@activeeon.com>