PORTAL-412 PORTAL-264 clientOnly add operation failed for datasource
Upon job submission the current behavior is to "manually" add the job in the list
to make it appear "right away" before the entire list of jobs is retrieved from
the server. The entire list of jobs is queried periodically.

My understanding is that this issue happens when the list of jobs with the new
submitted job is retrieved before the client displays it.

In that case, the error produced because the record with same id (job id) already exists.

I ignored this error in that case but the proper fix IMHO would probably to
avoid adding the job "manually" but force a refresh of the job's list upon job
submission.

However I find it risky to do before the release.

Change-Id: If0e92b53c05270367905ae158ec13d1837f7154f