Change node.jar description to use cred file.
The node.jar has rm.cred credentials inside, those can be outdated as soon as one changes the users. But the node.jar is created during the build process, meaning that the node.jar won't work anymore as soon as logins are changed.

Re-creation of the node.jar is not trivial, therefore we chose to not deliver credentials inside the node.jar and use user generated credentials.

A user can create credentials in the RM-Portal with two clicks.

So this change tells the user to create credentials and start the node.jar with those credentials.