Improve node limit message.
The node limit message was improved with this commit. It shows "Node Limit: 4" for a limit for 4 or "Node Limit: None" for no Limit.

A String in the attribute "count" threw an exception because of the formatting which is changed if the number is 0. The Integer.parse method threw an exception, it was changed to check for an equal content of "0".