Add placeholder for centerized logo (scheduler).
Add placeholder for centerized logo in the top bar. Default image is a
transparent square.

Once the war deployed, if you want to put a real image/logo, move it
under dist/war/scheduler/images and rename it extra-logo-center.png .

For instance (you can do it while the scheduler is running):
$ mv dist/war/scheduler/images/azure-cloud-transparent.png \
    dist/war/scheduler/images/extra-logo-center.png

And refresh the scheduler portal view.

For an optimal rendering, the logo should be something around 140x40 px.