Treeview render by deltas (#438)
TreeView in the RM portal re-render itself by snapshots of the state: S0, S1, .. SN. However, server now sends deltas and not snapshots, thus we have to make this view work with delta. For example, if we have already 10000 nodes, and server returns 100 nodes more, than we have to do 100 iteration and not 10100 iterations.